シーン名の定数を作成するクラスにEnumとテーブルを作成する処理を実装
