+++ Scripts/Editor/NameCreater/SceneNameTableCreator.cs.meta
@@ -1,8 +1,7 @@
 fileFormatVersion: 2
-guid: 1adf88db72be75341bf85244c75e0edb
-timeCreated: 1502761385
-licenseType: Free
+guid: 07f926789cf3c72498475be2fa062de2
 MonoImporter:
+  externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
