+++ Scripts/Utility/StringExtension.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b67610ef1276537439fa506120febf21
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
